tts/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-cli==0.0.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
inflect==7.3.1
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.60.0
numpy==2.0.1
packaging==24.1
Pillow==9.3.0
platformdirs==4.2.2
pooch==1.8.2
progressbar==2.5
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rotary-embedding-torch==0.6.4
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
starlette==0.38.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.13.4rc3
torch==2.5.0.dev20240817
torchaudio==2.4.0.dev20240817
torchvision==0.20.0.dev20240817
tortoise-tts @ git+https://github.com/neonbjb/tortoise-tts@2d6d3868f99d2fb97429741aba7d3a4314ae04aa
tqdm==4.66.5
transformers==4.31.0
typeguard==4.3.0
typer==0.12.4
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==12.0
zipp==3.20.0
